+++ src/imio/project/pst/setuphandlers.py
@@ -298,6 +298,36 @@ def _addPSTprojectspace(context):
          'key': "autres"},
     ]
     params['budget_types'] = budget_types
+    # datagridfield plan
+    plan_values = [
+        {'label': u"Agenda 21 local",
+         'key': "agenda-21-local"},
+        {'label': u"PCA Plan communal d'Aménagement",
+         'key': "pca-plan-communal-d-amenagement"},
+        {'label': u"Plan communal d'Urgence",
+         'key': "plan-communal-d-urgence"},
+        {'label': u"Plan communal de développement de la nature PCDN",
+         'key': "plan-communal-de-developpement-de-la-nature-pcdn"},
+        {'label': u"Plan communal de développement rural (PCDR)",
+         'key': "plan-communal-de-developpement-rural-pcdn"},
+        {'label': u"Plan d'ancrage communal",
+         'key': "plan-d-ancrage-communal"},
+        {'label': u"Plan de cohésion social (PCS)",
+         'key': "plan-de-cohesion-social-pcs"},
+        {'label': u"Plan de formation du personnel",
+         'key': "plan-de-formation-du-personnel"},
+        {'label': u"Plan de gestion",
+         'key': "plan-de-gestion"},
+        {'label': u"Plan de mobilité",
+         'key': "plan-de-mobilite"},
+        {'label': u"Plan global de prévention",
+         'key': "plan-global-de-prevention"},
+        {'label': u"Plan zonal de sécurité",
+         'key': "plan-zonal-de-securite"},
+        {'label': u"Schémas de développement commercial",
+         'key': "schemas-de-developpement-commercial"},
+    ]
+    params['plan_values'] = plan_values
     if getattr(site, '_TESTING_SITE_', False):
         params['budget_years'] = [2019, 2020, 2021, 2022, 2023, 2024]
     else:
@@ -506,26 +536,32 @@ def configure_pst(portal):
 
     if not registry.get('imio.project.settings.strategicobjective_fields'):
         registry['imio.project.settings.strategicobjective_fields'] = [
-            'IDublinCore.title', 'description_rich', 'reference_number', 'categories',
-            'IAnalyticBudget.projection', 'IAnalyticBudget.analytic_budget', 'budget', 'budget_comments',
+            'IDublinCore.title', 'description_rich', 'reference_number',
+            'categories', 'plan', 'IAnalyticBudget.projection',
+            'IAnalyticBudget.analytic_budget', 'budget', 'budget_comments',
             'observation', 'comments']
     if not registry.get('imio.project.settings.operationalobjective_fields'):
         registry['imio.project.settings.operationalobjective_fields'] = [
-            'IDublinCore.title', 'description_rich', 'reference_number', 'categories',
-            'result_indicator', 'priority', 'planned_end_date', 'representative_responsible',
-            'administrative_responsible', 'manager', 'extra_concerned_people', 'IAnalyticBudget.projection',
-            'IAnalyticBudget.analytic_budget', 'budget', 'budget_comments', 'ISustainableDevelopmentGoals.sdgs',
+            'IDublinCore.title', 'description_rich', 'reference_number',
+            'categories', 'plan', 'result_indicator', 'priority',
+            'planned_end_date', 'representative_responsible',
+            'administrative_responsible', 'manager', 'extra_concerned_people',
+            'IAnalyticBudget.projection', 'IAnalyticBudget.analytic_budget',
+            'budget', 'budget_comments', 'ISustainableDevelopmentGoals.sdgs',
             'observation', 'comments']
     if not registry.get('imio.project.settings.pstaction_fields'):
         registry['imio.project.settings.pstaction_fields'] = [
-            'IDublinCore.title', 'description_rich', 'reference_number', 'categories',
-            'result_indicator', 'planned_end_date', 'planned_begin_date', 'effective_begin_date',
-            'effective_end_date', 'progress', 'health_indicator', 'health_indicator_details',
-            'representative_responsible', 'manager', 'responsible', 'extra_concerned_people',
-            'IAnalyticBudget.projection', 'IAnalyticBudget.analytic_budget', 'budget', 'budget_comments',
+            'IDublinCore.title', 'description_rich', 'reference_number',
+            'categories', 'plan', 'result_indicator', 'planned_end_date',
+            'planned_begin_date', 'effective_begin_date', 'effective_end_date',
+            'progress', 'health_indicator', 'health_indicator_details',
+            'representative_responsible', 'manager', 'responsible',
+            'extra_concerned_people', 'IAnalyticBudget.projection',
+            'IAnalyticBudget.analytic_budget', 'budget', 'budget_comments',
             'ISustainableDevelopmentGoals.sdgs', 'observation', 'comments']
     if not registry.get('imio.project.settings.pstsubaction_fields'):
-        registry['imio.project.settings.pstsubaction_fields'] = registry['imio.project.settings.pstaction_fields']
+        registry['imio.project.settings.pstsubaction_fields'] = registry[
+                'imio.project.settings.pstaction_fields']
 
     if not registry.get('imio.project.settings.strategicobjective_budget_states'):
         registry['imio.project.settings.strategicobjective_budget_states'] = ['ongoing', 'achieved']